[WIP] feat: freehand draw lines on screen
Work in progress because line drawing behaviour is hardcoded.

What i mean by hardcoded is that the normally when the user first clicks
on the canvas, a new line is created and when they then drag their
cursor, the position of those drag events are added onto that line and
rendered to the user.

Right now, whenever the user first clicks down, a new line is created.
However when they drag their mouse, we add those mouse event positions
to only 1 line and the rest are left dangling.