fix: add mousemove points to the appropriate line
Previously, whenever the mouse was downed, a new line was created but
the points during the mouse drag events weren't added to that respective line.
It was added to only 1 particular line (which i hardcoded it to do so to
test if i was able to draw onto the screen).

This resulted in the behaviour where if I lifted the mouse and move away
and tried to draw again, the line would connect from where i previously
lifted off my mouse to where i mouse down-ed and dragged.

This fix now stores the id of any newly created line on mouse down and
when the user drags their cursor we find that line based on the stored
id and add the drag events' position points to that.